refactor: split existing component between steps
