+++ src/components/Deploy/1-DeployModeStep/index.tsx
@@ -1,3 +1,1 @@
-export function DeployModeStep() {
-  return <p>Deploy mode step</p>
-}
+export * from './DeployModeStep'
